style: clear the mechanical lint categories, and stop discarding export errors
434 -> 235 configured findings. Suite green at 600.

**RUF013, implicit Optional (26 -> 0).** `x: float = None` became `x: float | None = None`. A real
type defect, not cosmetics: the old signature says the parameter cannot be None while the default
is None.

**RUF046, unnecessary int() (45 -> 0).** All 45 were `int(round(x))` or `int(math.floor(x))` in
numpy-heavy detection code, which is exactly where this is worth checking rather than trusting: if
`round()` returned a numpy scalar instead of a Python int, the result would be used as an array
index and raise. Verified against every float type that reaches these sites -- float, float64,
float32, float16, int64 -- that `round()` and `math.floor()` each return a Python int equal to the
`int()`-wrapped form, and that both index and slice identically.

**RUF059, unused unpacked values (43 -> 3).** Genuine discards now say so with an underscore. One
was not a discard: `export/cli.py` unpacked `(success, collect_id, error)` in both its parallel and
sequential paths, counted the failures and threw the reason away -- so a failed export was reported
as a number and nothing else. Both paths now log which collect failed and why, which uses the
values rather than renaming them.

**TRY400, error-instead-of-exception (25 -> 0).** These now log the traceback. Checked against a
full 134-collect run log before applying: none of the 25 sites fires on a healthy run, so this
costs nothing in noise and only pays out when something actually breaks.

**RUF001-3, ambiguous unicode (74 -> 0)** via `allowed-confusables = ["×", "σ", "–", "−"]`, not by
switching the rules off. Each of the four is correct typography in this domain -- the multiplication
sign in binning factors and PSF multiples, sigma for standard deviation, an en dash in prose, a true
minus on negatives -- so the alternative was to spell them `x`, `sigma` and `-` and make the prose
worse. If you would rather have the ASCII, say so and I will swap them instead.

The three remaining RUF059 are tuple unpacks in CLI paths I have not decided on yet; the rest of the
235 is the tail below.

+++ senpai/engine/detection/point/fwhm.py
@@ -152,7 +152,7 @@ def measure_fwhm_from_catalog_stars(
     data = fits_image.data
     star_peaks: dict[int, float] = {}
     for star in valid_catalog_stars:
-        x0, y0 = int(round(star.x)), int(round(star.y))
+        x0, y0 = round(star.x), round(star.y)
         core = data[max(0, y0 - 2) : y0 + 3, max(0, x0 - 2) : x0 + 3]
         if core.size:
             star_peaks[id(star)] = float(core.max())
